+++ src/lib.rs
@@ -7,5 +7,7 @@ pub mod token;
 pub use filter::should_include_file;
 pub use git::get_git_diff;
 pub use path::{label, traverse_directory};
-pub use template::{extract_undefined_variables, handlebars_setup, render_template};
+pub use template::{
+    copy_to_clipboard, handle_undefined_variables, handlebars_setup, render_template, write_to_file,
+};
 pub use token::count_tokens;
